+++ triton-soccer-ai/src/main/java/com/triton/search/node1d/Node1d.java
@@ -3,9 +3,9 @@ package com.triton.search.node1d;
 import com.triton.search.base.GraphNode;
 
 public class Node1d implements GraphNode {
-    private final double value;
+    private final float value;
 
-    public Node1d(double value) {
+    public Node1d(float value) {
         this.value = value;
     }
