+++ triton-soccer-ai/src/main/java/com/triton/search/node2d/Euclidean2dScorer.java
@@ -4,7 +4,7 @@ import com.triton.search.base.Scorer;
 
 public class Euclidean2dScorer implements Scorer<Node2d> {
     @Override
-    public double computeCost(Node2d from, Node2d to) {
+    public float computeCost(Node2d from, Node2d to) {
         return from.getPos().dist(to.getPos()) + to.getPenalty();
     }
 }
