+++ triton-soccer-ai/src/main/java/com/triton/search/node2d/Manhattan2dScorer.java
@@ -4,7 +4,7 @@ import com.triton.search.base.Scorer;
 
 public class Manhattan2dScorer implements Scorer<Node2d> {
     @Override
-    public double computeCost(Node2d from, Node2d to) {
+    public float computeCost(Node2d from, Node2d to) {
         return Math.abs(to.getPos().x - from.getPos().x) + Math.abs(to.getPos().y - from.getPos().y);
     }
 }
